[TASK-04] Design patterns (fixes)
